+++ src/components/coauthor/GuideStepComp.tsx
@@ -9,7 +9,7 @@ import React, { useEffect, useState } from "react";
 import { useQuery } from "react-query";
 
 import { getArticleSteps } from "@/lib/coauthor";
-import { sendMessageToChatGPT } from "@/pages/api/openAI/helpers";
+import { sendMessageToChatGPT } from "@/lib/utils/utils";
 
 interface Step {
   name: string;
