+++ src/components/coauthor/ImproveComp.tsx
@@ -18,7 +18,7 @@ import {
 } from "firebase/firestore";
 import React, { useEffect, useState } from "react";
 
-import { sendMessageToChatGPT } from "@/pages/api/openAI/helpers";
+import { sendMessageToChatGPT } from "@/lib/utils/utils";
 
 import ImproveItemComp from "./ImproveItemComp";
