+++ src/components/coauthor/IssuesComp.tsx
@@ -4,7 +4,7 @@ import { doc, getFirestore, updateDoc } from "firebase/firestore";
 import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
 
 import { DESIGN_SYSTEM_COLORS } from "@/lib/theme/colors";
-import { sendMessageToChatGPT } from "@/pages/api/openAI/helpers";
+import { sendMessageToChatGPT } from "@/lib/utils/utils";
 
 interface Props {
   allContent: string;
