+++ src/lib/utils/utils.ts
@@ -15,6 +15,7 @@ import {
   TagsArray,
   TimeWindowOption,
 } from "../../knowledgeTypes";
+import { Post } from "../mapApi";
 
 export const isValidHttpUrl = (possibleUrl?: string) => {
   let url;
@@ -419,4 +420,22 @@ export const arraysAreEqual = (arr1: string[], arr2: string[]) => {
     }
   }
   return true;
+};
+
+export const sendMessageToChatGPT = async (messages: { role: string; content: string }[]) => {
+  const result: any = await Post("/openAI/request", {
+    messages,
+    model: "gpt-4o",
+    response_format: { type: "json_object" },
+  });
+  if (result?.response) {
+    try {
+      return JSON.parse(result?.response);
+    } catch (err: any) {
+      console.error(err.message);
+      return [];
+    }
+  } else {
+    return [];
+  }
 };
