+++ src/pages/api/openAI/helpers.ts
@@ -4,7 +4,6 @@ import { uploadFileToStorage } from "../STT";
 import { NextApiResponse } from "next";
 import { delay } from "@/lib/utils/utils";
 import { MODEL } from "@/lib/utils/constants";
-import { Post } from "@/lib/mapApi";
 
 const OpenAI = require("openai");
 
@@ -2684,24 +2683,6 @@ export async function callOpenAIChat(files: File[], userPrompt: string, systemPr
     throw error;
   }
 }
-
-export const sendMessageToChatGPT = async (messages: { role: string; content: string }[]) => {
-  const result: any = await Post("/openAI/request", {
-    messages,
-    model: "gpt-4o",
-    response_format: { type: "json_object" },
-  });
-  if (result?.response) {
-    try {
-      return JSON.parse(result?.response);
-    } catch (err: any) {
-      console.error(err.message);
-      return [];
-    }
-  } else {
-    return [];
-  }
-};
 
 export const generateImage = async (prompt: string) => {
   const response = await openai.images.generate({
